Revised Creating SnapShots
This document are instructions given by Computer Coder/YouTuber Virag-ky on how to create viable Commit statements, so this process can be conveyed to the "Real Coding World" and programmers to on how generate readable and comprehsive Commit statments.  

Git Statments were generated using the "Code With Mosh" exercises on Creating SnapShots.